refactor: split auth super-module into smaller modules
As this grows its become a big unweidly, split it out.
